Implement concurrent compute fragment
This commit implements a compute fragment that can retrieve
batches concurrently in the background. It takes in another fragment
and a buffer size, and creates a channel where it will pre-load
batches from its upstream using a goroutine.

This is not used anywhere yet, but I had the code already written out
so I thought we could merge it to main.