Fix data corruption during compaction
When reading rows from Parquet files, the library reuses underlying
memory buffers to save on allocations. This is by design and the user
is supposed to copy/clone rows if they are inteded to stick around for longer.

This commit modifies the compaction logic to copy rows when reading
sorted parts before feeding them into the merged row group. For now
there is no memory pooling in place, but it can be added later on
if we run into problems.