Reduce memory usage during compaction
The GenericBuffer from the segmentio library will allocate fixed sized
buffers for each column in the schema. Since our current schema is sparse
with most columns actually being empty, this leads to a high overhead
when creating sorted parts.

This commit replaces the GenericBuffer with a custom implementation
which maintains the sparse labels structure and avoids pre-allocating
memory ahead of time.